Add front-end signup
Added a button on the login screen to go to the sign-up page if needed.
Also added more state variables to control error messages and the header depending on whether or not the user is on the login/sign up page. Sign-up page also has password confirmation before sending a fetch request.